requirements.txt

# API e servidor
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Config e segurança
python-dotenv>=1.0.1
requests>=2.32.3
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0

# Banco de dados
sqlalchemy>=2.0.32
psycopg2-binary>=2.9.9

# Processamento de documentos
pypdf>=4.3.1
PyMuPDF>=1.24.9
python-docx>=1.1.2
python-pptx>=0.6.23
beautifulsoup4>=4.12.3
unstructured>=0.14.10

# LangChain + Ollama + Chroma
langchain>=0.2.10
langchain-community>=0.2.10
langchain-ollama>=0.1.0
langchain-chroma>=0.1.2
langchain-text-splitters>=0.2.2
chromadb>=0.5.0
